requirements.txt
google_genai==1.12.1 
python-dotenv==1.1.0